+++ library/src/com/viewpagerindicator/PageIndicator.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package com.jakewharton.android.view;
+package com.viewpagerindicator;
 
 import android.support.v4.view.ViewPager;