+++ sample/src/com/viewpagerindicator/SampleCirclesDefault.java
@@ -1,10 +1,10 @@
-package com.jakewharton.android.viewpagerindicator;
+package com.viewpagerindicator;
 
 import android.os.Bundle;
 import android.support.v4.view.ViewPager;
-import com.jakewharton.android.view.CirclePageIndicator;
+import com.viewpagerindicator.CirclePageIndicator;
 
-public class SampleCirclesInitialPage extends BaseSampleActivity {
+public class SampleCirclesDefault extends BaseSampleActivity {
 	@Override
 	protected void onCreate(Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
@@ -17,8 +17,5 @@ public class SampleCirclesInitialPage extends BaseSampleActivity {
 		
 		CirclePageIndicator indicator = (CirclePageIndicator)findViewById(R.id.indicator);
 		indicator.setViewPager(mPager);
-		indicator.setCurrentItem(mAdapter.getCount() - 1);
-		
-		//You can also do: indicator.setViewPager(pager, initialPage);
 	}
 }