+++ sample/src/com/viewpagerindicator/SampleCirclesInitialPage.java
@@ -1,14 +1,14 @@
-package com.jakewharton.android.viewpagerindicator;
+package com.viewpagerindicator;
 
 import android.os.Bundle;
 import android.support.v4.view.ViewPager;
-import com.jakewharton.android.view.CirclePageIndicator;
+import com.viewpagerindicator.CirclePageIndicator;
 
-public class SampleCirclesStyledLayout extends BaseSampleActivity {
+public class SampleCirclesInitialPage extends BaseSampleActivity {
 	@Override
 	protected void onCreate(Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
-		setContentView(R.layout.themed_circles);
+		setContentView(R.layout.simple_circles);
 		
 		mAdapter = new TestFragmentAdapter(getSupportFragmentManager());
 		
@@ -17,5 +17,8 @@ public class SampleCirclesStyledLayout extends BaseSampleActivity {
 		
 		CirclePageIndicator indicator = (CirclePageIndicator)findViewById(R.id.indicator);
 		indicator.setViewPager(mPager);
+		indicator.setCurrentItem(mAdapter.getCount() - 1);
+		
+		//You can also do: indicator.setViewPager(pager, initialPage);
 	}
 }