+++ sample/src/com/viewpagerindicator/SampleCirclesSnap.java
@@ -1,15 +1,13 @@
-package com.jakewharton.android.viewpagerindicator;
+package com.viewpagerindicator;
 
 import android.os.Bundle;
 import android.support.v4.view.ViewPager;
-import com.jakewharton.android.view.CirclePageIndicator;
+import com.viewpagerindicator.CirclePageIndicator;
 
-public class SampleCirclesStyledTheme extends BaseSampleActivity {
+public class SampleCirclesSnap extends BaseSampleActivity {
 	@Override
 	protected void onCreate(Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
-		
-		//The look of this sample is set via a style in the manifest
 		setContentView(R.layout.simple_circles);
 		
 		mAdapter = new TestFragmentAdapter(getSupportFragmentManager());
@@ -19,5 +17,6 @@ public class SampleCirclesStyledTheme extends BaseSampleActivity {
 		
 		CirclePageIndicator indicator = (CirclePageIndicator)findViewById(R.id.indicator);
 		indicator.setViewPager(mPager);
+		indicator.setSnap(true);
 	}
 }