+++ sample/src/com/viewpagerindicator/SampleCirclesStyledLayout.java
@@ -1,14 +1,14 @@
-package com.jakewharton.android.viewpagerindicator;
+package com.viewpagerindicator;
 
 import android.os.Bundle;
 import android.support.v4.view.ViewPager;
-import com.jakewharton.android.view.CirclePageIndicator;
+import com.viewpagerindicator.CirclePageIndicator;
 
-public class SampleCirclesSnap extends BaseSampleActivity {
+public class SampleCirclesStyledLayout extends BaseSampleActivity {
 	@Override
 	protected void onCreate(Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
-		setContentView(R.layout.simple_circles);
+		setContentView(R.layout.themed_circles);
 		
 		mAdapter = new TestFragmentAdapter(getSupportFragmentManager());
 		
@@ -17,6 +17,5 @@ public class SampleCirclesSnap extends BaseSampleActivity {
 		
 		CirclePageIndicator indicator = (CirclePageIndicator)findViewById(R.id.indicator);
 		indicator.setViewPager(mPager);
-		indicator.setSnap(true);
 	}
 }