+++ sample/src/com/viewpagerindicator/SampleCirclesStyledMethods.java
@@ -1,10 +1,10 @@
-package com.jakewharton.android.viewpagerindicator;
+package com.viewpagerindicator;
 
 import android.os.Bundle;
 import android.support.v4.view.ViewPager;
-import com.jakewharton.android.view.CirclePageIndicator;
+import com.viewpagerindicator.CirclePageIndicator;
 
-public class SampleCirclesDefault extends BaseSampleActivity {
+public class SampleCirclesStyledMethods extends BaseSampleActivity {
 	@Override
 	protected void onCreate(Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
@@ -17,5 +17,12 @@ public class SampleCirclesDefault extends BaseSampleActivity {
 		
 		CirclePageIndicator indicator = (CirclePageIndicator)findViewById(R.id.indicator);
 		indicator.setViewPager(mPager);
+		
+		final float density = getResources().getDisplayMetrics().density;
+		indicator.setBackgroundColor(0xFFCCCCCC);
+		indicator.setRadius(10 * density);
+		indicator.setFillColor(0xFF888888);
+		indicator.setStrokeColor(0xFF000000);
+		indicator.setStrokeWidth(2 * density);
 	}
 }