+++ sample/src/com/viewpagerindicator/SampleCirclesStyledTheme.java
@@ -1,13 +1,15 @@
-package com.jakewharton.android.viewpagerindicator;
+package com.viewpagerindicator;
 
 import android.os.Bundle;
 import android.support.v4.view.ViewPager;
-import com.jakewharton.android.view.CirclePageIndicator;
+import com.viewpagerindicator.CirclePageIndicator;
 
-public class SampleCirclesStyledMethods extends BaseSampleActivity {
+public class SampleCirclesStyledTheme extends BaseSampleActivity {
 	@Override
 	protected void onCreate(Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
+		
+		//The look of this sample is set via a style in the manifest
 		setContentView(R.layout.simple_circles);
 		
 		mAdapter = new TestFragmentAdapter(getSupportFragmentManager());
@@ -17,12 +19,5 @@ public class SampleCirclesStyledMethods extends BaseSampleActivity {
 		
 		CirclePageIndicator indicator = (CirclePageIndicator)findViewById(R.id.indicator);
 		indicator.setViewPager(mPager);
-		
-		final float density = getResources().getDisplayMetrics().density;
-		indicator.setBackgroundColor(0xFFCCCCCC);
-		indicator.setRadius(10 * density);
-		indicator.setFillColor(0xFF888888);
-		indicator.setStrokeColor(0xFF000000);
-		indicator.setStrokeWidth(2 * density);
 	}
 }