+++ sample/src/com/viewpagerindicator/SampleTitlesStyledMethods.java
@@ -1,9 +1,9 @@
-package com.jakewharton.android.viewpagerindicator;
+package com.viewpagerindicator;
 
 import android.os.Bundle;
 import android.support.v4.view.ViewPager;
-import com.jakewharton.android.view.TitlePageIndicator;
-import com.jakewharton.android.view.TitlePageIndicator.IndicatorStyle;
+import com.viewpagerindicator.TitlePageIndicator;
+import com.viewpagerindicator.TitlePageIndicator.IndicatorStyle;
 
 public class SampleTitlesStyledMethods extends BaseSampleActivity {
 	@Override