+++ sample/src/com/viewpagerindicator/SampleTitlesUnderline.java
@@ -1,10 +1,11 @@
-package com.jakewharton.android.viewpagerindicator;
+package com.viewpagerindicator;
 
 import android.os.Bundle;
 import android.support.v4.view.ViewPager;
-import com.jakewharton.android.view.TitlePageIndicator;
+import com.viewpagerindicator.TitlePageIndicator;
+import com.viewpagerindicator.TitlePageIndicator.IndicatorStyle;
 
-public class SampleTitlesInitialPage extends BaseSampleActivity {
+public class SampleTitlesUnderline extends BaseSampleActivity {
 	@Override
 	protected void onCreate(Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
@@ -17,8 +18,6 @@ public class SampleTitlesInitialPage extends BaseSampleActivity {
 		
 		TitlePageIndicator indicator = (TitlePageIndicator)findViewById(R.id.indicator);
 		indicator.setViewPager(mPager);
-		indicator.setCurrentItem(mAdapter.getCount() - 1);
-		
-		//You can also do: indicator.setViewPager(pager, initialPage);
+		indicator.setFooterIndicatorStyle(IndicatorStyle.Underline);
 	}
 }