+++ sample/src/com/viewpagerindicator/TestFragmentAdapter.java
@@ -1,4 +1,4 @@
-package com.jakewharton.android.viewpagerindicator;
+package com.viewpagerindicator;
 
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;